[VisualStudio] avoid ambiguous call to overloaded isnan method

+++ src/plugins/legacy/meshMapping/meshMapping.cpp
@@ -78,7 +78,7 @@ public:
             
             img->TransformPhysicalPointToIndex(itkPoint, itkIndex);
             typename ImageType::PixelType pixelValue = img->GetPixel(itkIndex);
-            if (!std::isnan(pixelValue))
+            if (!std::isnan(static_cast<double>(pixelValue)))
             {
                 interpolatedValues->SetValue(i, static_cast<double>(pixelValue));
             }
